requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
Werkzeug==3.0.1
bcrypt==4.1.2
PyJWT==2.8.0
uuid==1.30